Refactor HomePage and localizes IAP service
Move the initialization of the `IAPService` from the `SplashScreen` to the `HomePage`'s state.
This change localizes the `IAPService` dependency, reducing prop drilling and making pages like `SignInPage` and `SettingsPageController` independent of it.
A loading state is introduced for the `HomePage` to wait for IAP initialization.
Additionally, the `HomePage`'s build method is extensively refactored into smaller, more focused helper widgets and methods, greatly improving code readability and maintainability.